Updated
1.Removed visualization in 3_SEIR.R, as this is now accomplished through
a seprate function in modelvis_wip.R .
2.Also added summary variables so visualization will be easier.
3.modelvis_wip.R now working.

In the future, it might be nice to have input for the axis and title
labels, but this looks like it will have to be some sort of direct
change to the output JSON. Not dealing with that now. It is as efficient
as I know how to make it with my current knowledge, and better than
storing locals into the global environment to call with subset() which
was what I was doing.