Diagnostic
Tau-leaping seems to be screwing with the introduction of new cases at any given time. Solving with Gillespie seems to work fine, but introduction method is either not intuitive to the algorithm being used or is fundamentally flawed. I'm certain there is a better way to introduce, but I haven't found it yet. I'll keep trying