Change a-plenty
- Benchmark re-tested and confirmed on my terrible HP 4 core laptop.
Quite astounding, the computation time is on the order of a square-root
better over 4 cores.
- Even after setting the seed, multicore optimized doesn't seem to yield
the same results as the non-multicore optimized one. There is almost
always fewer observations with the multicore-optimized one, but I'm not
sure if that's significant.
- Reworked matrix subsetting for my sanity. From 5 lines to 2, and much
more readable.
- Attempted to tidy code to conform to google code standards and make a
bit more readible, but we will see.
- Split mul_ins function from the main batch_plot function because I
feel like the nested functions bit is getting to be a little much. May
not keep, but it makes more sense to me to call it as a global function
than mulitple local ones.