Library
- Croots function works well, but has issues running on school computers
where g++ cannot be set to path. Built a package which can be
run instead.
- Some tweaks to Epi_detect to error handle sourcing the C++ function,
and additional output if verbose is turned on.
- Removed tau leaping limits from multicore, as these are no longer
needed.
- I'm working on better argument handling with multicore (multicore_2)
to avoid the amount of arguments which must be typed in each time. I
will test a few methods for the best balance of performance and
usability.