Edits
- `data` updated to reflect some more liberal estimates of population
levels in Sweden.
- `Epi_detect` has now been updated with error handling in the case
where outbreaks extended past simulation length and analysis resulted in
error. It will now produce a warning message and skip the iteration, but
it would be wise to re-run the sim if this occurs in practice.
- `multicore` : rearranged some assignment and error handling to allow
testing for the case where the insertion given is out of bounds (If you
can't make the 200 pound gorilla between the chair and the keyboard
smarter, at least make the function smarter, amirite?).
- `trial.cpp` : I had wanted to used unsigned ints to handle a few of
the values to avoid the possibility of negative or excessively large
values, but this of course doesn't play nicely with SEXPs which munch on
doubles. Who needs available RAM space anyways?
- I did something in the Excel file as well. Who knows at this point.